requirements.txt
astroid==2.2.5
aws-shell==0.2.1
awscli==1.16.185
backports.csv==1.0.7
boto3==1.9.167
botocore==1.12.175
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
configobj==5.0.6
coverage==4.5.3
defusedxml==0.5.0
des==1.0.5
diff-match-patch==20181111
Django==2.1.7
django-admin-tools==0.8.1
django-allauth==0.39.1
django-cors-headers==2.4.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-nose==1.4.6
django-rest-auth==0.9.3
django-rest-passwordreset==0.9.7
django-reversion==3.0.3
django-test-without-migrations==0.6
django-xadmin==0.5.0
djangorestframework==3.9.2
docutils==0.14
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
isort==4.3.20
jdcal==1.4
jmespath==0.9.4
lazy-object-proxy==1.4.1
libsass==0.18.0
mccabe==0.6.1
nose==1.3.7
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==5.4.1
prompt-toolkit==1.0.16
psycopg2==2.7.1
pyasn1==0.4.5
pycrypto==2.6.1
Pygments==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
sendgrid==5.6.0
simplejson==3.16.0
six==1.12.0
tablib==0.13.0
typed-ast==1.4.0
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
ua-parser==0.8.0
user-agents==2.0
django-user-agents==0.4.0
django-ckeditor==5.7.1 
pyDes==2.0.1
spyne==2.13.2a0
